tomcat/src/main/java/org/apache/catalina/session/Session.java
package org.apache.catalina.session;

import java.util.HashMap;
import java.util.Map;

public class Session {

    private final String id;
    private final Map<String, Object> attributes;

    public Session(final String id) {
        this.id = id;
        this.attributes = new HashMap<>();
    }

    public String getId() {
        return id;
    }

    public Object getAttribute(final String name) {
        return attributes.get(name);
    }

    public void setAttribute(final String name, final Object attribute) {
        attributes.put(name, attribute);
    }

    public void removeAttribute(final String name) {
        attributes.remove(name);
    }
}
